// src/app/shared/interfaces/search.interface.ts
export class ISearchParams {
    nmea_d_bgs_d: number[];
    nmea_d_heel_d: number[];
    nmea_d_rwa_d: number[];
    nmea_d_tws_d: number[];
    start_time: number [ ];
    State: string [];
    Channel: string;
    anomaly_score_d:  number;
    start: number;
    count: number;
    type: string;
    ascendant:string;
}